refactor(#742): relocate receiver-type method dispatch to call-receiver-method.ts
Wave B chunk 4, step 2. Move compileReceiverMethodCall (the receiverType-keyed
tail of the property-access arm: user-class methods, Number/BigInt/Boolean
wrapper methods, generators, typed arrays, valueOf/toString) out of calls.ts
into a new sibling module src/codegen/expressions/call-receiver-method.ts. 19
calls.ts internals exported to it (incl. the two STANDALONE_TA_* sets); calls.ts
imports compileReceiverMethodCall back (lazy cycle).

calls.ts: 12,564 -> 9,534 LOC. compileCallExpression: ~6,480 -> ~3,430 LOC. New
call-receiver-method.ts: 3,165 LOC. This empties the property-access arm.

Pure relocation — emitted Wasm byte-identical (prove-emit-identity IDENTICAL,
39/39). tsc 0. Smoke test extended with user-class method + Number toFixed/
toString(radix) cases. loc-budget + coercion-sites (11 relocated sites, net-zero)
pass with allowances in the #742 frontmatter; oracle-ratchet net-zero.

✓
