+++ plan/issues/742-extract-and-refactor-compilecallexpression-3.md
@@ -16,9 +16,11 @@ loc-budget-allow:
   - src/codegen/expressions/call-identifier.ts
   - src/codegen/expressions/call-builtin-static.ts
   - src/codegen/expressions/call-namespace-static.ts
+  - src/codegen/expressions/call-receiver-method.ts
 coercion-sites-allow:
   - src/codegen/expressions/call-identifier.ts
   - src/codegen/expressions/call-namespace-static.ts
+  - src/codegen/expressions/call-receiver-method.ts
 # 2026-07-12 (#3182 groom): elevated medium→high. The EXTRACTION half is done
 # (calls.ts exists, 18,753 LOC — see #803, closed as landed); the live scope is
 # the REFACTOR half: break up compileCallExpression inside
@@ -230,3 +232,42 @@ wrapper methods, generators, typed arrays, valueOf/toString, ~8705–end of arm)
 that section DOES use the arm-level `receiverType`/`receiverClassName` locals, so
 the next slice threads those in (or lifts the `receiverType = …` computation into
 the helper). Issue stays `in-progress`.
+
+## Progress — Wave B chunk 4: receiver-type method dispatch (2026-07-14, sendev-waveb)
+
+Empties the property-access arm. Extracts its **receiver-type method dispatch**
+tail — the `receiverType`-keyed half (user-class instance methods, Number /
+BigInt / Boolean wrapper methods, generator methods, typed-array methods, and
+the generic valueOf / toString / toLocaleString fallbacks), calls.ts 6788–9835
+post-chunk-3 (~3,048 LOC, from the `let receiverType = …` line to the arm close)
+— into a new sibling module `src/codegen/expressions/call-receiver-method.ts`
+(`compileReceiverMethodCall`).
+
+The block was lifted **from the `receiverType = …` computation onward**, so the
+helper computes `receiverType` / `receiverClassName` / `recvTsType` itself (all
+arm-local, unused after the arm) — self-contained on
+`ctx`/`fctx`/`expr`/`propAccess`/`expectedType` (`expectedType` is threaded
+through; its single use is one argument site). The one `return undefined` match
+in the span is a comment, not a statement. Same two verified steps; 19 `calls.ts`
+internals exported to it (17 helper functions + the two `STANDALONE_TA_*` typed-
+array HOF sets; `compileCallExpression` / `BUILTIN_CLASS_NAMES` /
+`emitWrapperDynamicMethodCall` were already exported).
+
+**Result**: `compileCallExpression` ~6,480 → ~3,430 LOC; `calls.ts` 12,564 →
+9,534 LOC; new `call-receiver-method.ts` 3,165 LOC.
+
+**Proof**: `prove-emit-identity` IDENTICAL across all 39 `(file,target)` emits
+after each step; `tsc` 0 (two module-scope `STANDALONE_TA_*` consts the import
+scan missed were caught by tsc and exported); `oracle-ratchet` net-zero;
+`loc-budget` + `coercion-sites` (11 relocated coercion sites, net-zero) pass with
+allowances added above. Smoke test extended with user-class method + Number
+`toFixed`/`toString(radix)` cases (wasm≡JS).
+
+**With the property-access arm now fully decomposed**, `compileCallExpression` is
+a lean dispatch skeleton: the guard prelude, the property-access arm (now just
+regexp guards + calls into `compileBuiltinStaticCall` /
+`compileNamespaceStaticCall` / `compileReceiverMethodCall`), the identifier
+dispatch (`compileIdentifierCall`), and the remaining tail arms — IIFE, super
+(`compileSuperMethodCall`), element-access, call-of-call, and conditional callee
+(~3,430 LOC total). Those tail arms are the remaining slices. Issue stays
+`in-progress`.
