+++ tests/issue-742.test.ts
@@ -179,3 +179,33 @@ describe("#742 namespace static dispatch (compileNamespaceStaticCall)", () => {
     ]);
   });
 });
+
+// Slice 4 (#742 Wave B): the receiver-type method dispatch — the receiverType-
+// keyed tail of compileCallExpression's property-access arm (user-class instance
+// methods, Number wrapper methods, valueOf/toString) — was moved verbatim into
+// the sibling module call-receiver-method.ts (compileReceiverMethodCall). These
+// tests pin the behaviour of those moved method-call paths (wasm ≡ JS).
+describe("#742 receiver-type method dispatch (compileReceiverMethodCall)", () => {
+  it("user-class instance method dispatch", async () => {
+    await assertEquivalent(
+      `class Adder { base: number; constructor(b: number) { this.base = b; } add(x: number): number { return this.base + x; } twice(x: number): number { return this.add(x) + this.add(x); } }
+       export function m1(): number { const a = new Adder(10); return a.add(5); }
+       export function m2(): number { const a = new Adder(3); return a.twice(4); }`,
+      [
+        { fn: "m1", args: [] },
+        { fn: "m2", args: [] },
+      ],
+    );
+  });
+
+  it("Number wrapper methods: toFixed / toString(radix)", async () => {
+    await assertEquivalent(
+      `export function nfix(): number { return (3.14159).toFixed(2) === "3.14" ? 1 : 0; }
+       export function nhex(): number { return (255).toString(16) === "ff" ? 1 : 0; }`,
+      [
+        { fn: "nfix", args: [] },
+        { fn: "nhex", args: [] },
+      ],
+    );
+  });
+});
